chore: fix more merge conflicts
